Update: time calculation function
